Properly structuring files, updated code, refreshed memory on code in preparation to write V2
